+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 ### Changed
 
 - #2867 - Make the Versioned Clientlibs transformer pick up css link tags without a type attribute if  the attribute rel="stylesheet" is set
+- #2865 - Reports - Turn absolute property paths to relative to prevent report breakage when malformed (abs path) data is assed 
 
 ### Fixed
 
@@ -23,7 +24,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 - #2712 - MCP Content Fragment Import: Improve import of Date and DateTime fields
 - #2869 - Support 500 error pages on AEM CS using x-aem-error-pass = true HTTP response header
 - #2857 - Fixed issue with Marketo integration loading marketo form
-- 
+
 ## 5.3.0 - 2022-04-15
 
 ### Fixed